Remove SGD2ModL exports
The SGD2ModL exports are not needed for a DLL that can act standalone.
