+++ SGD2FreeDisplayFix/src/dll_main.c
@@ -45,6 +45,10 @@
 
 #include <windows.h>
 
+#include "patches/patches.h"
+
+static struct Patches patches;
+
 BOOL WINAPI DllMain(
     HINSTANCE hinstDLL,
     DWORD fdwReason,
@@ -52,13 +56,15 @@ BOOL WINAPI DllMain(
 ) {
   switch (fdwReason) {
     case DLL_PROCESS_ATTACH: {
-      Sgd2fml_Mod_OnLoadMpqs();
+      patches = Patches_Init();
+      Patches_Apply(&patches);
 
       return TRUE;
     }
 
     case DLL_PROCESS_DETACH: {
-      Sgd2fml_Mod_OnUnloadMpqs();
+      Patches_Remove(&patches);
+      Patches_Deinit(&patches);
 
       return TRUE;
     }
